image_uploader/requirements.txt
ensemble-boxes
Flask
Flask-MQTT
Flask-UUID
numpy
Pillow
PyYAML
requests
psutil
